feat: use slash to separate rpc id
